[RW-3070][risk=low] Moves search input to new file, fixes linter errors, uses colors module.
